Create Given the row number n. Print the n-th row of Pascal’s triangle.
The steps are as follows:

We will use a loop(say c) to iterate over each column i.e. from 1 to n. And for each column, we will do the following steps:
First, we will consider n-1 as n and c-1 as r.
After that, we will simply calculate the value of the combination using a loop. 
The loop will run from 0 to r. And in each iteration, we will multiply (n-i) with the result and divide the result by (i+1).
Finally, we will print the element.
Finally, the entire row will be printe